Started refactoring
HAS PROBLEMS: check_login is not found within captcha.py :(
Added 'try-except' to handle database transaction failures when calling Token.get_valid()
